+++ new/memory.cpp
@@ -39,13 +39,13 @@ Memory::~Memory()
    safe malloc 
 ------------------------------------------------------------------------- */
 
-void *Memory::smalloc(int n, const char *name)
+void *Memory::smalloc(size_t n, const char *name)
 {
   if (n == 0) return NULL;
   void *ptr = malloc(n);
   if (ptr == NULL) {
     char str[128];
-    sprintf(str,"Failed to allocate %d bytes for array %s",n,name);
+    sprintf(str,"Failed to allocate %lu bytes for array %s",n,name);
     error->one(str);
   }
   return ptr;
@@ -55,7 +55,7 @@ void *Memory::smalloc(int n, const char *name)
    safe malloc with alignment
 ------------------------------------------------------------------------- */
 
-void *Memory::smalloc_align(int n, int nalign, const char *name)
+void *Memory::smalloc_align(size_t n, int nalign, const char *name)
 {
   if (n == 0) return NULL;
   //void *ptr;
@@ -64,7 +64,7 @@ void *Memory::smalloc_align(int n, int nalign, const char *name)
   void *ptr = malloc(n);
   if (ptr == NULL) {
     char str[128];
-    sprintf(str,"Failed to allocate %d bytes for array %s",n,name);
+    sprintf(str,"Failed to allocate %lu bytes for array %s",n,name);
     error->one(str);
   }
   return ptr;
@@ -84,7 +84,7 @@ void Memory::sfree(void *ptr)
    safe realloc 
 ------------------------------------------------------------------------- */
 
-void *Memory::srealloc(void *ptr, int n, const char *name)
+void *Memory::srealloc(void *ptr, size_t n, const char *name)
 {
   if (n == 0) {
     sfree(ptr);
@@ -94,7 +94,7 @@ void *Memory::srealloc(void *ptr, int n, const char *name)
   ptr = realloc(ptr,n);
   if (ptr == NULL) {
     char str[128];
-    sprintf(str,"Failed to reallocate %d bytes for array %s",n,name);
+    sprintf(str,"Failed to reallocate %lu bytes for array %s",n,name);
     error->one(str);
   }
   return ptr;
